Rename args.svelte.ts to args.ts and update imports to remove .svelte extension
